Django/pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "snow-sr", email = "joaofelipi877@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
